requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
beautifulsoup4==4.8.2
bs4==0.0.1
certifi==2019.11.28
chardet==3.0.4
html5lib==1.0.1
idna==2.9
multidict==4.7.5
python-dotenv==0.12.0
requests==2.23.0
six==1.14.0
slackclient==2.5.0
soupsieve==2.0
urllib3==1.25.8
webencodings==0.5.1
yarl==1.4.2